refactor(agents): extract candidate-adoption lifecycle from lifecycle_ops
Third carve of the lifecycle_ops.rs god-file (the most fix-churned file in the
codebase). Extracts the candidate-adoption state machine — the entry point of a
Cook attempt — into lifecycle_candidate_adoption.rs: project_cook_alias_adoption,
start_candidate_adoption(_with_rerun_policy), start_candidate_adoption_gate,
heartbeat/checkpoint/candidate_adoption_cancel_requested, finish_candidate_adoption,
reconcile_candidate_adoption, plus candidate_adoption_recovery_outcome and its
no_runner_job_recorded/is_pre_provider_transport_recovery helpers.

Behavior-preserving pure move following the agent_task_lifecycle module pattern
(mod X; pub use X::*). New module opens with use super::* and is re-exported through
mod.rs, so external agent_task_lifecycle:: paths resolve unchanged. Widened
project_cook_alias_adoption and reconcile_candidate_adoption to pub(crate) for their
remaining lifecycle_ops (status) callers, and EXPIRED_LAB_HANDOFF_REASON to pub(crate)
for the recovery-outcome reconstruction.

lifecycle_ops.rs: 3708 -> 3342 LOC.
